Updated reset logic so that ASYC_REG gets applied properly.
Vivado was doing stupid shit again. With the ASYNC_REG property now being set in this VHDL module rather than in the XDC, it was putting a LUT in between each register stage. Exactly what it's not supposed to do. Moving the reset into the sequential block seems to fix it and allow dumbass Vivado to infer the right thing.



git-svn-id: file:///afs/slac/g/reseng/svn/repos@12452 0848cbd7-7569-0410-ab05-8cc0a0c67708