Updated reset logic so that ASYC_REG gets applied properly.
Vivado was doing stupid shit again. With the ASYNC_REG property now being set in this VHDL module rather than in the XDC, it was putting a LUT in between each register stage. Exactly what it's not supposed to do. Moving the reset into the sequential block seems to fix it and allow dumbass Vivado to infer the right thing.



git-svn-id: file:///afs/slac/g/reseng/svn/repos@12452 0848cbd7-7569-0410-ab05-8cc0a0c67708

Former-commit-id: 402a30c2c63ad1c78413468efec2ad5bc91971e8
Former-commit-id: 4aa3abaf8689dd9b8fa262900bfa4545eea1715b
Former-commit-id: 0add42d471df6d733fd0a8d797eaead85b1ecd92